refactor(created utils elements in src/utils): refactored all cn, baseurl, metadata in all files
