+++ package.json
@@ -76,6 +76,7 @@
     "@trpc/client": "next",
     "@trpc/react-query": "next",
     "@trpc/server": "next",
+    "@types/color": "^3.0.6",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "@upstash/redis": "^1.34.3",
@@ -83,6 +84,7 @@
     "@vercel/speed-insights": "^1.0.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "color": "^4.2.3",
     "drizzle-orm": "^0.31.4",
     "drizzle-zod": "^0.5.1",
     "framer-motion": "^11.11.9",
@@ -99,6 +101,7 @@
     "react-hook-form": "^7.53.0",
     "resend": "^3.5.0",
     "server-only": "^0.0.1",
+    "slugify": "^1.6.6",
     "sonner": "^1.5.0",
     "superjson": "^2.2.1",
     "tailwind-merge": "^2.5.4",
