+++ src/app/(dashboard)/app/_components/active-button.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 import { Button } from '@/primitives/button';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
