+++ src/app/(site)/blog/[slug]/page.tsx
@@ -2,7 +2,9 @@ import { getBlogPosts } from '@/lib/mdx';
 import { notFound } from 'next/navigation';
 import { CustomMDX, MDXComponents } from '../../_components/custom-mdx';
 import type { Metadata } from 'next';
-import { constructMetadata, getBaseUrl } from '@/lib/utils';
+import { constructMetadata } from '@/utils/construct-metadata';
+import { getBaseUrl } from '@/utils/base-url';
+
 import { constants } from '@/constants';
 
 export const dynamic = 'force-static';
