+++ src/primitives/popover.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import * as PopoverPrimitive from '@radix-ui/react-popover';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 const Popover = PopoverPrimitive.Root;
