+++ src/primitives/select.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from 'lucide-react';
 import * as SelectPrimitive from '@radix-ui/react-select';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 const Select = SelectPrimitive.Root;
